Hi-proj 2. MORO/Code/Backend - MORO/requirements.txt
python-dotenv>=1.0.0
Flask>=2.3.0
flask-cors>=4.0.0
openai>=1.40.0
notion-client>=2.2.1
APScheduler>=3.10.4
gunicorn